move Profile.js from component to views
